Refactor package name from kotlin_android_sample to kotlinandroidsample
