+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Kotlin_android_sample"
+        android:theme="@style/Theme.KotlinAndroidSample"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
