+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">kotlin_android_sample</string>
+    <string name="app_name">Kotlin Android Sample</string>
 </resources>
