+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.Kotlin_android_sample" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.KotlinAndroidSample" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.Kotlin_android_sample" parent="Base.Theme.Kotlin_android_sample" />
+    <style name="Theme.KotlinAndroidSample" parent="Base.Theme.KotlinAndroidSample" />
 </resources>
